Fixed the orphan code checkFavoriteStatus()

+++ src/main/java/com/directi/train/tweetapp/services/FeedStore.java
@@ -3,7 +3,6 @@ package com.directi.train.tweetapp.services;
 import com.directi.train.tweetapp.model.FeedItem;
 import com.directi.train.tweetapp.model.UserProfileItem;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcTemplate;
 import org.springframework.stereotype.Service;
@@ -99,12 +98,20 @@ public class FeedStore {
         return oldFeedItems;
     }
 
+    public boolean isFavorited(Long tweetId, Long userId) {
+        return db.queryForInt(String.format("select count(*) from favorites where tweet_id = %d and user_id = %d", tweetId, userId)) > 0;
+    }
+
     public boolean favoriteTweet(Long creatorId, Long tweetId, Long userId) {
+        if (isFavorited(tweetId, userId))
+            return false;
         return db.update(String.format("insert into favorites (tweet_id, user_id) values (%d, %d)", tweetId, userId)) > 0;
     }
 
     public boolean unFavoriteTweet(Long creatorId, Long tweetId, Long userId) {
-        return db.update(String.format("delete from favorites where tweet_id = %d and user_id = %d", tweetId, userId)) > 0;
+        if (isFavorited(tweetId, userId))
+            return db.update(String.format("delete from favorites where tweet_id = %d and user_id = %d", tweetId, userId)) > 0;
+        return false;
     }
 
     public FeedItem reTweet(Long creatorId, Long tweetId, Long userId) {
